Wild modifications! I finally learned how to create a branch for experimental versions.
# App UI/UX
- Halfmoon Elegant toggles now use Font Awesome’s moon SVG (official Free 7.1.0 path) on every /app_ui page, so the theme button visually matches the design language.
- Added “Halfmoon CSS framework (MIT)” credit + link in README → Credits to stay compliant.
- each page loads halfmoon@2.0.2/css/halfmoon.min.css + css/cores/halfmoon.elegant.css, sets <html data-bs-core="elegant">, auto-detects OS theme via JS, and exposes the new button (moon icon) to flip data-bs-theme between dark/light.
-  now honors the landing-page preference—redirect via HOME_REDIRECT_URL when set or render the bundled/template override when present.

# refactor: split app.py into ghost_gateway package
- Cleared the unused ghost_gateway/ stub folder so we can do a clean module split when we tackle the app.py refactor.
- Added ghost_gateway.create_app() factory and kept HTTPS enforcement/DB bootstrap there.
- Ported every admin route (login, referrers, tokens, dashboard, logs, CSV) into ghost_gateway/routes/admin.py with shared utilities and proper rate limiting.
- Moved public landing + /read/<token> logic into ghost_gateway/routes/public.py, preserving access logging and Ghost fetch behavior.
- Simplified app.py to just start the factory; updated auth redirect helper to target the new blueprint namespace.
