+++ app_ui/login.html
@@ -1,111 +1,86 @@
 <!DOCTYPE html>
-<html lang="en">
+<html lang="en" data-bs-core="elegant">
 <head>
     <meta charset="utf-8">
     <title>{{ site.title or "FriendLink Gateway" }} · Admin Login</title>
     <meta name="viewport" content="width=device-width, initial-scale=1">
+    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/halfmoon@2.0.2/css/halfmoon.min.css">
+    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/halfmoon@2.0.2/css/cores/halfmoon.elegant.css">
     <style>
-        :root {
-            color-scheme: light dark;
-        }
         body {
-            font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", sans-serif;
-            background: #0f172a;
-            color: #f8fafc;
-            margin: 0;
             min-height: 100vh;
-            display: flex;
-            align-items: center;
-            justify-content: center;
-            padding: 2rem;
         }
-        .card {
-            background: rgba(15, 23, 42, 0.9);
-            border-radius: 16px;
-            padding: 2rem;
+        .login-card {
             width: min(420px, 100%);
-            box-shadow: 0 30px 80px rgba(15, 23, 42, 0.6);
         }
-        h1 {
-            margin: 0 0 1.5rem 0;
-            font-size: 1.6rem;
-        }
-        label {
-            display: block;
-            margin-bottom: 1rem;
-        }
-        label span {
-            display: block;
-            font-size: 0.9rem;
-            margin-bottom: 0.4rem;
-            color: #cbd5f5;
-        }
-        input[type="text"],
-        input[type="password"] {
-            width: 100%;
-            padding: 0.65rem 0.8rem;
-            border-radius: 10px;
-            border: 1px solid rgba(148, 163, 184, 0.4);
-            background: rgba(15, 23, 42, 0.6);
-            color: #f8fafc;
-        }
-        button {
-            width: 100%;
-            padding: 0.85rem;
-            border-radius: 10px;
-            border: none;
-            background: linear-gradient(135deg, #2563eb, #7c3aed);
-            color: #fff;
-            font-size: 1rem;
-            font-weight: 600;
-            cursor: pointer;
-            margin-top: 0.5rem;
-        }
-        .error {
-            background: rgba(248, 113, 113, 0.15);
-            border: 1px solid rgba(248, 113, 113, 0.4);
-            color: #fecaca;
-            padding: 0.65rem 0.8rem;
-            border-radius: 10px;
-            margin-bottom: 1rem;
-        }
-        .meta {
-            font-size: 0.85rem;
-            color: #94a3b8;
-            margin-top: 1rem;
-            text-align: center;
-        }
-        @media (max-width: 520px) {
-            body {
-                padding: 1.5rem 1rem;
-            }
-            .card {
-                padding: 1.5rem;
-            }
+        .theme-toggle {
+            position: fixed;
+            top: 1rem;
+            right: 1rem;
+            z-index: 1000;
         }
     </style>
+    <script>
+        function updateTheme() {
+            const prefersDark = window.matchMedia('(prefers-color-scheme: dark)').matches;
+            document.documentElement.setAttribute('data-bs-theme', prefersDark ? 'dark' : 'light');
+        }
+        function toggleTheme() {
+            const html = document.documentElement;
+            const current = html.getAttribute('data-bs-theme');
+            html.setAttribute('data-bs-theme', current === 'dark' ? 'light' : 'dark');
+        }
+        updateTheme();
+        window.matchMedia('(prefers-color-scheme: dark)').addEventListener('change', updateTheme);
+    </script>
 </head>
-<body>
-    <main class="card">
-        <h1>{{ site.title or "FriendLink Gateway" }}</h1>
-        {% if error %}
-        <div class="error">{{ error }}</div>
-        {% endif %}
-        <form method="post" action="{{ url_for('admin_login', next=next_url) }}">
-            <label>
-                <span>Username</span>
-                <input type="text" name="username" autofocus required value="{{ username or '' }}">
-            </label>
-            <label>
-                <span>Password</span>
-                <input type="password" name="password" required>
-            </label>
-            {% if next_url %}
-            <input type="hidden" name="next" value="{{ next_url }}">
-            {% endif %}
-            <button type="submit">Sign in</button>
-        </form>
-        <p class="meta">Secure session login · rotates every {{ ADMIN_SESSION_HOURS or 12 }} hours.</p>
-    </main>
+<body class="with-custom-webkit-scrollbars with-custom-css-scrollbars">
+    <button class="btn btn-secondary btn-sm theme-toggle" onclick="toggleTheme()" aria-label="Toggle theme">
+        <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 640 640" fill="currentColor" width="18" height="18">
+            <path d="M303.3 112.7C196.2 121.2 112 210.8 112 320C112 434.9 205.1 528 320 528C353.3 528 384.7 520.2 412.6 506.3C309.2 482.9 232 390.5 232 280C232 214.2 259.4 154.9 303.3 112.7zM64 320C64 178.6 178.6 64 320 64C339.4 64 358.4 66.2 376.7 70.3C386.6 72.5 394 80.8 395.2 90.8C396.4 100.8 391.2 110.6 382.1 115.2C321.5 145.4 280 207.9 280 280C280 381.6 362.4 464 464 464C469 464 473.9 463.8 478.8 463.4C488.9 462.6 498.4 468.2 502.6 477.5C506.8 486.8 504.6 497.6 497.3 504.6C451.3 548.8 388.8 576 320 576C178.6 576 64 461.4 64 320z"/>
+        </svg>
+    </button>
+    <div class="page-wrapper">
+        <div class="content-wrapper d-flex align-items-center justify-content-center py-40 px-20">
+            <div class="card shadow-lg login-card p-25">
+                <h1 class="font-size-22 mb-20">{{ site.title or "FriendLink Gateway" }}</h1>
+                {% if error %}
+                <div class="alert alert-danger">
+                    {{ error }}
+                </div>
+                {% endif %}
+                <form method="post" action="{{ url_for('admin_login', next=next_url) }}" class="w-full">
+                    <div class="form-group">
+                        <label for="username">Username</label>
+                        <input type="text"
+                               id="username"
+                               name="username"
+                               class="form-control"
+                               required
+                               autofocus
+                               value="{{ username or '' }}">
+                    </div>
+                    <div class="form-group">
+                        <label for="password">Password</label>
+                        <input type="password"
+                               id="password"
+                               name="password"
+                               class="form-control"
+                               required>
+                    </div>
+                    {% if next_url %}
+                    <input type="hidden" name="next" value="{{ next_url }}">
+                    {% endif %}
+                    <button type="submit" class="btn btn-primary btn-block">
+                        Sign in
+                    </button>
+                </form>
+                <p class="text-muted font-size-12 mt-20 mb-0 text-center">
+                    Sessions rotate every {{ ADMIN_SESSION_HOURS or 12 }} hours · CSRF protected
+                </p>
+            </div>
+        </div>
+    </div>
+    <script src="https://cdn.jsdelivr.net/npm/halfmoon/js/halfmoon.min.js" defer></script>
 </body>
 </html>
