+++ env.template
@@ -6,6 +6,7 @@ GHOST_CONTENT_API_KEY=
 APP_BASE_URL=https://your-gateway-domain.com
 DEFAULT_REDIRECT=https://your-ghost-blog.com/#/portal/signup
 CUSTOM_CSS_URL=
+HOME_REDIRECT_URL=
 SITE_TITLE=
 SITE_DESCRIPTION=
 SITE_ICON_URL=
